# Optimize `TokenMap` updates and traversals for modified files only