--- a/sway-lsp/src/core/session.rs
+++ b/sway-lsp/src/core/session.rs
@@ -27,6 +27,7 @@ use pkg::{
 };
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use std::{
+    collections::BTreeMap,
     path::PathBuf,
     sync::{atomic::AtomicBool, Arc},
     time::SystemTime,
@@ -298,13 +299,31 @@ pub fn traverse(
     results: Vec<(Option<Programs>, Handler)>,
     engines_clone: &Engines,
     session: Arc<Session>,
+    lsp_mode: Option<LspConfig>,
 ) -> Result<Option<CompileResults>, LanguageServerError> {
     let _p = tracing::trace_span!("traverse").entered();
-    session.token_map.clear();
+
+    let modified_file = lsp_mode.and_then(|mode| {
+        mode.file_versions
+            .iter()
+            .find_map(|(path, version)| version.map(|_| path.clone()))
+    });
+    if let Some(path) = &modified_file {
+        eprintln!("Clearing tokens for file: {:?}", path);
+        session.token_map.remove_tokens_for_file(path);
+    } else {
+        eprintln!("Clearing all tokens");
+        session.token_map.clear();
+    }
+
+    // session.token_map.clear();
+    // JOSH TODO: we probably only want to clear the metrics for the modified file also.
     session.metrics.clear();
+
     let mut diagnostics: CompileResults = (Vec::default(), Vec::default());
     let results_len = results.len();
     for (i, (value, handler)) in results.into_iter().enumerate() {
+        dbg!();
         // We can convert these destructured elements to a Vec<Diagnostic> later on.
         let current_diagnostics = handler.consume();
         diagnostics = current_diagnostics;
@@ -337,7 +356,22 @@ pub fn traverse(
             let path = engines.se().get_path(source_id);
             let program_id = program_id_from_path(&path, engines)?;
             session.metrics.insert(program_id, metrics);
+
+            eprintln!("path: {:?}", path);
+            if let Some(modified_file) = &modified_file {
+                let modified_program_id = program_id_from_path(&modified_file, engines)?;
+
+                eprintln!(
+                    "program_id: {:?} | modified_program_id: {:?}",
+                    program_id, modified_program_id
+                );
+
+                if program_id != modified_program_id {
+                    continue;
+                }
+            }
         }
+        dbg!();
 
         // Get a reference to the typed program AST.
         let typed_program = typed
@@ -355,32 +389,42 @@ pub fn traverse(
 
         // The final element in the results is the main program.
         if i == results_len - 1 {
+            let now = std::time::Instant::now();
             // First, populate our token_map with sway keywords.
             lexed_tree::parse(&lexed, &ctx);
+            eprintln!("Lexed took {:?}", now.elapsed());
 
+            let now = std::time::Instant::now();
             // Next, populate our token_map with un-typed yet parsed ast nodes.
             let parsed_tree = ParsedTree::new(&ctx);
             parsed_tree.collect_module_spans(&parsed);
-            parse_ast_to_tokens(&parsed, &ctx, |an, _ctx| parsed_tree.traverse_node(an));
+            parse_ast_to_tokens(&parsed, &ctx, &modified_file, |an, _ctx| {
+                parsed_tree.traverse_node(an)
+            });
+            eprintln!("Parsed took {:?}", now.elapsed());
 
+            let now = std::time::Instant::now();
             // Finally, populate our token_map with typed ast nodes.
             let typed_tree = TypedTree::new(&ctx);
             typed_tree.collect_module_spans(typed_program);
-            parse_ast_to_typed_tokens(typed_program, &ctx, |node, _ctx| {
+            parse_ast_to_typed_tokens(typed_program, &ctx, &modified_file, |node, _ctx| {
                 typed_tree.traverse_node(node);
             });
+            eprintln!("Typed took {:?}", now.elapsed());
 
+            let now = std::time::Instant::now();
             let compiled_program = &mut *session.compiled_program.write();
             compiled_program.lexed = Some(lexed);
             compiled_program.parsed = Some(parsed);
             compiled_program.typed = Some(typed_program.clone());
+            eprintln!("Write compiled_program took {:?}", now.elapsed());
         } else {
             // Collect tokens from dependencies and the standard library prelude.
-            parse_ast_to_tokens(&parsed, &ctx, |an, ctx| {
+            parse_ast_to_tokens(&parsed, &ctx, &modified_file, |an, ctx| {
                 dependency::collect_parsed_declaration(an, ctx);
             });
 
-            parse_ast_to_typed_tokens(typed_program, &ctx, |node, ctx| {
+            parse_ast_to_typed_tokens(typed_program, &ctx, &modified_file, |node, ctx| {
                 dependency::collect_typed_declaration(node, ctx);
             });
         }
@@ -413,7 +457,9 @@ pub fn parse_project(
         return Err(LanguageServerError::ProgramsIsNone);
     }
 
-    let diagnostics = traverse(results, engines, session.clone())?;
+    let now = std::time::Instant::now();
+    let diagnostics = traverse(results, engines, session.clone(), lsp_mode.clone())?;
+    eprintln!("Traverse took {:?}", now.elapsed());
     if let Some(config) = &lsp_mode {
         // Only write the diagnostics results on didSave or didOpen.
         if !config.optimized_build {
@@ -435,9 +481,21 @@ pub fn parse_project(
 fn parse_ast_to_tokens(
     parse_program: &ParseProgram,
     ctx: &ParseContext,
+    modified_file: &Option<PathBuf>,
     f: impl Fn(&AstNode, &ParseContext) + Sync,
 ) {
-    let nodes = parse_program
+    let should_process = |node: &&AstNode| {
+        modified_file
+            .as_ref()
+            .map(|path| {
+                node.span
+                    .source_id()
+                    .map_or(false, |id| ctx.engines.se().get_path(id) == *path)
+            })
+            .unwrap_or(true)
+    };
+
+    parse_program
         .root
         .tree
         .root_nodes
@@ -448,17 +506,31 @@ fn parse_ast_to_tokens(
                 .submodules_recursive()
                 .flat_map(|(_, submodule)| &submodule.module.tree.root_nodes),
         )
-        .collect::<Vec<_>>();
-    nodes.par_iter().for_each(|n| f(n, ctx));
+        .filter(should_process)
+        .collect::<Vec<_>>()
+        .par_iter()
+        .for_each(|n| f(n, ctx));
 }
 
 /// Parse the [ty::TyProgram] AST to populate the [TokenMap] with typed AST nodes.
 fn parse_ast_to_typed_tokens(
     typed_program: &ty::TyProgram,
     ctx: &ParseContext,
+    modified_file: &Option<PathBuf>,
     f: impl Fn(&ty::TyAstNode, &ParseContext) + Sync,
 ) {
-    let nodes = typed_program
+    let should_process = |node: &&ty::TyAstNode| {
+        modified_file
+            .as_ref()
+            .map(|path| {
+                node.span
+                    .source_id()
+                    .map_or(false, |id| ctx.engines.se().get_path(id) == *path)
+            })
+            .unwrap_or(true)
+    };
+
+    typed_program
         .root
         .all_nodes
         .iter()
@@ -468,8 +540,10 @@ fn parse_ast_to_typed_tokens(
                 .submodules_recursive()
                 .flat_map(|(_, submodule)| &submodule.module.all_nodes),
         )
-        .collect::<Vec<_>>();
-    nodes.par_iter().for_each(|n| f(n, ctx));
+        .filter(should_process)
+        .collect::<Vec<_>>()
+        .par_iter()
+        .for_each(|n| f(n, ctx));
 }
 
 /// Create runnables if the `TyProgramKind` of the `TyProgram` is a script.