--- a/sway-lsp/src/traverse/lexed_tree.rs
+++ b/sway-lsp/src/traverse/lexed_tree.rs
@@ -4,11 +4,16 @@ use crate::{
 };
 use rayon::iter::{ParallelBridge, ParallelIterator};
 use sway_ast::{
-    assignable::ElementAccess, attribute::Annotated, expr::LoopControlFlow, ty::TyTupleDescriptor, Assignable, CodeBlockContents, ConfigurableField, Expr, ExprArrayDescriptor, ExprStructField, ExprTupleDescriptor, FnArg, FnArgs, FnSignature, IfCondition, IfExpr, ItemAbi, ItemConfigurable, ItemConst, ItemEnum, ItemFn, ItemImpl, ItemImplItem, ItemKind, ItemStorage, ItemStruct, ItemTrait, ItemTypeAlias, ItemUse, MatchBranchKind, ModuleKind, Pattern, PatternStructField, Statement, StatementLet, StorageEntry, StorageField, TraitType, Ty, TypeField, UseTree
+    assignable::ElementAccess, attribute::Annotated, expr::LoopControlFlow, ty::TyTupleDescriptor,
+    Assignable, CodeBlockContents, ConfigurableField, Expr, ExprArrayDescriptor, ExprStructField,
+    ExprTupleDescriptor, FnArg, FnArgs, FnSignature, IfCondition, IfExpr, ItemAbi,
+    ItemConfigurable, ItemConst, ItemEnum, ItemFn, ItemImpl, ItemImplItem, ItemKind, ItemStorage,
+    ItemStruct, ItemTrait, ItemTypeAlias, ItemUse, MatchBranchKind, ModuleKind, Pattern,
+    PatternStructField, Statement, StatementLet, StorageEntry, StorageField, TraitType, Ty,
+    TypeField, UseTree,
 };
 use sway_core::language::{lexed::LexedProgram, HasSubmodules};
 use sway_types::{Ident, Span, Spanned};
-
 
 pub struct LexedTree<'a> {
     ctx: &'a ParseContext<'a>,
@@ -23,10 +28,7 @@ impl<'a> LexedTree<'a> {
         node.value.parse(self.ctx);
     }
 
-    pub fn collect_module_kinds(
-        &self,
-        lexed_program: &LexedProgram,
-    ) {
+    pub fn collect_module_kinds(&self, lexed_program: &LexedProgram) {
         insert_module_kind(self.ctx, &lexed_program.root.tree.kind);
         lexed_program
             .root
@@ -35,25 +37,7 @@ impl<'a> LexedTree<'a> {
                 insert_module_kind(self.ctx, &dep.module.tree.kind);
             });
     }
-
 }
-
-// pub fn parse(lexed_program: &LexedProgram, ctx: &ParseContext) {
-//     insert_module_kind(ctx, &lexed_program.root.tree.kind);
-//     adaptive_iter(&lexed_program.root.tree.items, |item| {
-//         item.value.parse(ctx);
-//     });
-
-//     lexed_program
-//         .root
-//         .submodules_recursive()
-//         .for_each(|(_, dep)| {
-//             insert_module_kind(ctx, &dep.module.tree.kind);
-//             adaptive_iter(&dep.module.tree.items, |item| {
-//                 item.value.parse(ctx);
-//             });
-//         });
-// }
 
 fn insert_module_kind(ctx: &ParseContext, kind: &ModuleKind) {
     match kind {